The rounded corners of the header result column have not been removed; now I have tried adding a small icon to indicate that it is draggable.
Change the drag button to use an SVG icon and make it fill the entire cell.

More pronounced color contrast and shadow effectsThe rounded corners of the header result column have not been removed; now I have tried adding a small icon to indicate that it is draggable. Change the drag button to use an SVG icon and make it fill the entire cell.